# openapi.yaml
openapi: 3.0.3
info:
  title: Neuro Resume API
  description: |
    API для приложения создания резюме с помощью AI-интервью.
    Приложение проводит интерактивное интервью с пользователем,
    собирает информацию о его опыте и навыках, а затем генерирует
    профессиональное резюме.
  version: 1.0.0
  contact:
    name: Neuro Resume Team
    email: support@neuroresume.com
servers:
  - url: https://api.neuroresume.com/v1
    description: Production server
  - url: https://staging-api.neuroresume.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
paths:
...
  /interview/sessions:
    get:
      tags:
        - interview
      summary: Получить список сессий интервью
      description: Возвращает все интервью-сессии текущего пользователя
      operationId: getInterviewSessions
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Фильтр по статусу сессии
          schema:
            type: string
            enum: [in_progress, completed, abandoned]
        - name: limit
          in: query
          description: Количество записей на странице
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Смещение для пагинации
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Список сессий успешно получен
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InterviewSession'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - interview
      summary: Создать новую сессию интервью
      description: Начать новое интервью для создания резюме
      operationId: createInterviewSession
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  description: Язык интервью
                  enum: [ru, en]
                  default: ru
                  example: ru
      responses:
        '201':
          description: Сессия успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /interview/sessions/{sessionId}:
    get:
      tags:
        - interview
      summary: Получить информацию о сессии
      description: Детальная информация об интервью-сессии
      operationId: getInterviewSession
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID сессии интервью
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о сессии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterviewSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - interview
      summary: Удалить сессию интервью
      description: Удаление сессии и всех связанных данных
      operationId: deleteInterviewSession
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID сессии интервью
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Сессия успешно удалена
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /interview/sessions/{sessionId}/messages:
    get:
      tags:
        - interview
      summary: Получить историю сообщений
      description: Все сообщения в рамках интервью-сессии
      operationId: getSessionMessages
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID сессии интервью
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: История сообщений
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    format: uuid
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - interview
      summary: Отправить сообщение в интервью
      description: Отправка ответа пользователя и получение следующего вопроса от AI
      operationId: sendMessage
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID сессии интервью
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Текст ответа пользователя
                  example: Я работал разработчиком в компании X более 3 лет
                  minLength: 1
                  maxLength: 5000
      responses:
        '200':
          description: Сообщение отправлено, получен ответ AI
          content:
            application/json:
              schema:
                type: object
                properties:
                  userMessage:
                    $ref: '#/components/schemas/Message'
                  aiResponse:
                    $ref: '#/components/schemas/Message'
                  progress:
                    type: object
                    description: Прогресс заполнения резюме
                    properties:
                      percentage:
                        type: integer
                        minimum: 0
                        maximum: 100
                        example: 45
                      completedSections:
                        type: array
                        items:
                          type: string
                        example: [personal_info, work_experience]
                      currentSection:
                        type: string
                        example: skills
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /interview/sessions/{sessionId}/complete:
    post:
      tags:
        - interview
      summary: Завершить интервью
      description: Пометить интервью как завершенное и начать генерацию резюме
      operationId: completeInterview
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          description: ID сессии интервью
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Интервью завершено
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    $ref: '#/components/schemas/InterviewSession'
                  resumeId:
                    type: string
                    format: uuid
                    description: ID созданного резюме
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Интервью уже завершено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный при авторизации
  schemas:
...
    InterviewSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 456e7890-e89b-12d3-a456-426614174111
        userId:
          type: string
          format: uuid
        status:
          type: string
          enum: [in_progress, completed, abandoned]
          example: in_progress
        language:
          type: string
          enum: [ru, en]
          example: ru
        progress:
          type: object
          properties:
            percentage:
              type: integer
              minimum: 0
              maximum: 100
              example: 60
            completedSections:
              type: array
              items:
                type: string
              example: [personal_info, work_experience, education]
            currentSection:
              type: string
              example: skills
        messageCount:
          type: integer
          example: 15
        createdAt:
          type: string
          format: date-time
          example: 2025-10-12T09:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2025-10-12T10:30:00Z
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: null
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 789e0123-e89b-12d3-a456-426614174222
        sessionId:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, ai]
          example: ai
        content:
          type: string
          example: Расскажите, пожалуйста, о вашем последнем месте работы
        metadata:
          type: object
          nullable: true
          description: Дополнительные данные сообщения
          properties:
            extractedData:
              type: object
              description: Извлеченные из ответа данные для резюме
        createdAt:
          type: string
          format: date-time
          example: 2025-10-12T10:15:00Z
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Общее количество записей
          example: 150
        limit:
          type: integer
          description: Количество записей на странице
          example: 20
        offset:
          type: integer
          description: Текущее смещение
          example: 0
        hasMore:
          type: boolean
          description: Есть ли еще записи
          example: true
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation failed
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: Invalid email format
  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: email
                  message: Invalid email format
                - field: password
                  message: Password must be at least 6 characters